+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Change Log
 
+## [Unreleased]
+
+### Breaking Changes
+- `print-only` option on pages changed to use new `include-in` option
+
+### Added
+- `include-in` YAML option available for both group and page definitions to allow inclusion of groups and/or pages in specific targets. Valid values are `all` (default), `print`, and `module`.
+
+### Fixed
+- Fixed broken drop caps lettering
+
 ## 1.0.14
 
 ### Added